adds insert_money class to dry code
