+++ item.rb
@@ -1,0 +1,3 @@
+class Item
+
+end
