+++ purchase.rb
@@ -1,0 +1,2 @@
+def Purchase
+end
